requirements.txt
asgiref~=3.8
Django~=5.2
sqlparse==0.5.0
python-decouple==3.8
django-debug-toolbar~=5.2.0
django-admin-interface
django-colorfield
djangorestframework
drf-spectacular
